pyproject.toml
[tool.poetry]
name = "github-analysis"
version = "0.1.0"
description = "GitHub PR Analysis Tool"
authors = ["Kelvin Davis <archidit7@gmail.com>"]
packages = [
    { include = "github_analysis", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.38"}
aiohttp = "^3.11.12"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
qdrant-client = "^1.13.2"
openai = "^1.61.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dev-dependencies]
black = "^25.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^6.0.0"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
